requirements.txt
line-bot-sdk==2.4.2
Flask==2.2.5